requirements.txt
aiohttp==3.9.1
aiomultiprocess==0.9.0
colorama==0.4.6
matplotlib==3.8.3
numpy==1.24.3
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.1.0
Pillow==10.3.0
Requests==2.31.0
retina_face==0.0.17
retinaface_pytorch==0.0.7
